Get column or property for use in mapping

This function returns a an expression to get a specified column from a dataset (or a property from a layer).

Usage

get_column(column)

Arguments

column

The name of the column or property to get.

Value

A list representing the expression to get the column.