Get queried features from a map as an sf object

This function retrieves the results of a feature query triggered by query_features(). It returns the features as a deduplicated sf object.

Usage

get_queried_features(map)

Arguments

map

A map object (mapboxgl, maplibre) or proxy object (mapboxgl_proxy, maplibre_proxy, mapboxgl_compare_proxy, maplibre_compare_proxy)

Value

An sf object containing the queried features, or an empty sf object if no features were found

Examples

if (FALSE) { # \dontrun{
# In a Shiny server function:
observeEvent(input$query_button, {
    proxy <- maplibre_proxy("map")
    query_features(proxy, layer_id = "counties")
    features <- get_queried_features(proxy)
    print(nrow(features))
})
} # }